+++ lib/src/sensors.dart
@@ -1,0 +1,9 @@
+part of flutter_sensors;
+
+class Sensors {
+  static const int ACCELEROMETER = 1;
+  static const int GYROSCOPE = 4;
+  static const int MAGNETIC_FIELD = 2;
+  static const int LINEAR_ACCELERATION = 10;
+  static const int STEP_COUNTER = 19;
+}