+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 ## 0.0.1
 
 * Function to register a listener to sensor changes.
-* Function to unregister a previously registered listener.
+* Function to un-register a previously registered listener.
 * Added callback that inform changes in a sensor with the accuracy of the reading.